Complete rewrite of parser to be block based
This rewrite of the parser is much more cleaner. It uses a block tree as
intermediate representation instead of parsing directly to the HTML
tree.

I think it's feature complete regarding the old functionality while also
implementing the following:

 - Numbered lists
 - List nesting
 - Consecutive lines that started with tags now show up. Before there
   was a bug where only the first one would show up.

It also reuses the new inline content parser for paragraphs and
headings. Before we had duplicated code for handling these blocks.