Move base directory to the user's home
I've been storing note source data inside the project's directory, this
isn't right, it should be stored in the user's home directory. This way
there is a clearer split about what is the user's data, and what gets
generated.

Created a testing directory where I will add sample notes. This shoul be
the start of some unit testing.

Fixed an issue with the copy_changed() method where it wouldn't work
when copying from a relative to an absolute path. Don't strip the
leading '/'.