Implement user tags and use the API to implement math tag procesing
Because we use KaTeX to render math expressions. It's only available in
Javascript so this commit also adds the Duktape Javascript engine. I've
noticed calling into Javascript is quite slow so the number of
javascript dependecise should be kept to a minimum. To the point where
I would like to reimplement all these dependencies in C and avoid having
the Javascript engine at all.

Because calling into Javascript is slow, I've also added a toggle to the
build that disables it and replaces these library functions with a stub.
I also hope this added friction of calling into Javascript (need to
implement function and a strub) will also reduce the usage of the
Javascript engine.