Abstract out parsing of note titles
We want to build an initial map of titles, for this we need to parse
them from the runtime. Because this may be rewuired during parsing
itself we can't wait for all notes to be parsed, we need to perform a
partial pass to get the titles. This is a step towards that.

Also implemented a proper error mesage implementation for the parser. We
now get better errors when running the tests.