Ignore line breaks in link URLs, improve white space tests
This moves linebreak suppression from the block parser into the inline
parser. In the weird case of a URL that was broken into multiple lines I
think this is good, because the default behavior of replacing newline
with space is always wrong. In this case we know the correct behavior is
to remove the newline character.

Multiple White spaces are still consumed at the block parsing level.
Improved the testing for this, so that it also checks tab characters are
considered spaces. Checked they are also ignored at the block level.
While doing this I also improved the printing of debug strings that make
visible these white space characters and clearly show if the string is
terminated in a linebreak or not.