Compute default margin at the correct time
There was a bug where a list followed by a paragraph starting with a tag
would nest this paragraph inside the list block (but outside the last
list item block).

This happened because margin was set to the column after some matches
were made, in particular the scanner had already advanced over the \
character of the tag, which resulted in margin being 1 not 0 as
expected.

Added a test for this case and made tests fail on parsing warnings too.