Implement automatic computation of notes
This changes the way notes are processed in a significant way. Added
link computation for all notes, which happens before executing any
callback. This allows computing orphan notes from a callback because we
know all likns have already been computed.

It also adds an example of a custom tag that creates a string in psplx
format which gets parsed into blocks.