Add negative testing to test suite
Big refactoring of the end user parsing API, because it's mostly a proof
of concept at this point I'm now using the control flow reuse. This way
I don't have to spend time updating API signatures and I can reuse a
base implementation of how a single note goes through the  processing
pipeline.

This then allow reusing the same source of the code when implementing
the runtime that reads all notes. This runtime doesn't process a single
note until the end, intead it applies the same stage to all notes, then
moves on to the next stage. This is required because we need to make
sure for example that links are resolvable during parsing.

I'm also starting to abstract out the testing framework so I can also
create programmatic tests, and not just file based ones.