Refactor the old expanded notation to be valid Turtle
With the addition of floating elements to the cannonical syntax, it's
not valid Turtle anymore. Still, TSPLX data can be mapped in a
straightforward way to RDF in Turtle notation. The result is actually
close to what I was calling expanded notation before so I have
refactored that code and renamed it to convey that it should generate
valid .ttl files.