Don't embed nodes as lists
We can't really treat nodes as list nodes. This is because a a single
node can belong to multiple lists (it can be the subject of multiple
triples, or be a floating value as well as a subject of a triple).

Added a more complex test case that forces some cases where the old
implementation would fail.