Improvements to file utilities
When parsing a canonical file name, return an object instead of a tuple.
This makes code more readable and future proof than just using the
indices of each member.

Extend the caonical file format to support a location section. This
allows having multiple files corresponding to multiple pages of the same
document. They will have the same id but increasing location. This
location allows for sections, for example a big poster tha has rows and
is double sided will have file names with 3 levels in the location
section.

Implement an API to simplify the sequential creation of documents that
use the canonical file name format. This makes it easy to scan multi
page documents. This simplifies working with the new location format
for canonical file names.

Split file utility tests into a separate file.