Add support for Markdown's inline code syntax using `
Now ` must be escaped as \` in paragraphs if the desired result is the actual
` character. Same thing when the ` character is necessary inside a code
inline block that uses Markdown's notation.

Spaces are stripped when using this notation and to end a code string
with \ a space must be added between the last \ and the closing `.