Implement image embedding by ID
This is starting to implement ID based functionality. This mekes it
possible to link a file from a note by just using the ID which in
principle is much more flexible. I think later we could extend this to
link files that are outside of weaver's data folder (like pictures or
files for books and papers). Also, starting to use identifiers more
broadly will make it possible in the future to easily attach metadata to
files.

Also, I changed the binary tree API to remove the _tree prefix in
generated functions.