Fix escaping of LaTeX expressions
I wasn't correctly expanding " characters which conflicted with the
literal string that was being used to build the string parameter passed
to KaTeX through Javascript.

Also, set notation was broken because I was escaping \{ and \} as PSPLX
operators inside the tag content. We now don't do any operator escaping
when tag content termination is based on the expectation of balanced
braces.