Keep the HTML tree for each node
This will allow to make modifications to it after the note parsing but
before actual rendering. This will be useful for example to applend HTML
sections to rendered pages.

Also, make all resources of notes be allocated in a note pool instead of
upstream in the runtime's pool. Right now we are not handling note
destruction, which is okay because we don't delete notes at runtime.