Add linking from inline data tags
Inside paragraphs it's now posible to use notation ^person[John Doe]
to make a link to a data entity of type person with name "John Doe".
This correctly creates link and backlink references so orphan note
computation will work for them.

This breaks one canonical test serialization because we're now rendering
these in nested notation automatically. Should improve canonical
notation to remove this automatic magic based on the presence of a root
node. Instead we should have "flat mode" and "nested note" that does
require a root.

Also, added a common function that sets a string_t from a sstring_t.
Looked like such a common operation that doing so would reduce the size
of the code a bit.