Support data tags at the beginning of a line
This is the start of a big refatoring on the parser architecture. The
idea is to actually reuse some of the parsing constructs and make them
stateless and not require any memory applocations.

This allows reusing the tag parsing constructs to make decisions early
on, about if something is a valid tag, and other characteristics like,
is it at the end of a block, inline in a paragraph, does it have
parameters or content?.