Begin implementation of public site generation
This adds configuration parameters to set the title notes in the public
version as well as specify which entity types should be skipped.

I reworked the whole title note configuration reading and stopped using
Python templates for this. Instead use Mustache from C. This removes the
Python side handling of this, so it's not necessary in the cache
anymore. I also removed the helper methods in pymk.py as they didn't
seem to be that useful. I also removed all note graph building at the
Python level because now we can do all of this from within C.