Implement automatic redacting when publishing
In the public version of the notes, it's not good to create links to
hidden notes that will be broken. If the link is contained between () we
redact out the whole parenthesis from the output. If it's not, then we
warn that we're creating a broken link.

When listing entities with \entity_list make an actual link between
the note to the entity, such that the backlink also shows. To make this
work and still have a functional Orphans note, I had to add sorting to
the callback handling so that \orphan_list gets handled last.

Improved the web rendering for mobile devices.