Progress consolidating \note tag handling as a case of generic text handling
Tests seem to be in a working state but there are lots of failures.
Generation of my full notes is completely broken.

Remove the negative test for unlinked note error message. We now allow
this in all cases and just link to a virtual entity.

I still need some way to reduce typo mistakes when linking. Autocomplete
for entities would be ideal, a warning about similarli named entities
would be probably a good enough intermediate step?.