Implement fragment navigation
When linking, it's very common to want to point to a specific section so
the browser scrolls directly to it. This adds IDs to headings and an
event listener to handle the hash portion of an URL.

Also fixed a crash on public notes when creating entity lists.

Add a function to common.h that turns a string into lowercase.