Begin implementation of web server, improve tests
Add a simple plaveholder for a Flask web application together with the
necessary scripts to deploy it as a wheel.

Fix all tests so all headings include the new "id" attribute that will
be used for fragment links. We now special ASCII characters by removing
them from the title when forming the ID.

Fixed the Python test harness to fully support nested tests.

Consolidate all tests into a single snip.

Add a new kind of test that generates a full site (instead of a single
page).