Implement automated deployment and testing for Python server
This adds an implementation of an autolink service to the server. It
fully automates the API testing and deployment process into production.

There are scripts to locally test in virtual environments.

This also enables static site generation testing for the full site and
public one. It will make much simpler to test features related to the
Weaver site as a whole and not just the TSPLX/PSPLX parsers in
isolation.

It starts moving into a JSON format for web data instead of JavaScript
with embedded variable definitions. This makes the client side code be
closer to how it would look like for a REST API.